refactor: move inline functions to separate file
