+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/MutableMultiSetImpl.kt
@@ -316,28 +316,6 @@ internal class MutableMultiSetImpl<E> : MutableMultiSet<E> {
 //    }
 
     /**
-     * Accumulates value starting with [initial] value and applying [operation] from left to right
-     * to current accumulator value and each element.
-     *
-     * Returns the specified [initial] value if the collection is empty.
-     *
-     * @param initial [T]: initial value for applying operation
-     * @param [operation] (T, E) -> T: function that takes current accumulator value and an element, and calculates the next accumulator value.
-     * @return [T]: the accumulated value, or [initial] if the MultiSet is empty
-     */
-//    override fun <T> fold(initial: T, operation: (T, E) -> T): T {
-//        var acc = initial
-//
-//        countsMap.forEach {
-//            val value = it.key
-//            val count = it.value
-//            repeat(count) { acc = operation(acc, value) }
-//        }
-//
-//        return acc
-//    }
-
-    /**
      * Update the values of [list] and [string] to match the current values in the set.
      */
     private fun updateList() {
